requirements.txt
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-macaddress==1.8.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Jinja2==3.1.2
openapi-codec==1.3.2
packaging==21.3
Pillow==9.3.0
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
simplejson==3.17.6
psycopg2
psycopg2-binary==2.9.5
PyYAML==6.0
django-import-export
drf-yasg
newrelic
pycryptodome==3.16.0
pycryptodomex==3.16.0
geopy==2.2.0
python-dateutil==1.4
geocoder==1.38.1
django-environ
#new comment